Working bezier isoline rendering
started with line -> quad geometry shader

+++ shaders/rect/single/vert.glsl
@@ -3,7 +3,7 @@ layout(location = 0) in vec2 aPos;
 
 uniform vec2 pos;
 uniform vec2 size;
-uniform ivec2 screen_size;
+uniform vec2 screen_size;
 
 void main() {
     vec2 converted_pos = pos / screen_size * 2 - 1;
